requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
prometheus_flask_exporter==0.22.4
gunicorn==20.1.0